+++ src/hooks/useModal.js
@@ -1,0 +1,14 @@
+import { useCallback, useState } from 'react';
+
+const useModal = () => {
+  const [showModal, setShowModal] = useState(false);
+  const onCloseModal = useCallback(() => {
+    setShowModal(false);
+  }, []);
+  const openModal = useCallback(() => {
+    setShowModal((prev) => !prev);
+  }, []);
+  return [showModal, onCloseModal, openModal];
+};
+
+export default useModal;
